+++ core-ng/src/main/java/core/framework/internal/web/HTTPHandler.java
@@ -24,8 +24,6 @@ import io.undertow.server.HttpHandler;
 import io.undertow.server.HttpServerExchange;
 import io.undertow.util.HeaderMap;
 import io.undertow.util.HttpString;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
@@ -49,7 +47,6 @@ public class HTTPHandler implements HttpHandler {
 
     public final RateControl rateControl = new RateControl(1000);   // save at max 1000 group/ip combination
 
-    private final Logger logger = LoggerFactory.getLogger(HTTPHandler.class);
     private final LogManager logManager;
     private final SessionManager sessionManager;
     private final ResponseHandler responseHandler;